Work in progress streamlining job submission
The original combination of generate_scripts.py and
mkmaster_scripts.py is really ugly.  This is partly
completed work to make it more elegant and general.

Basically, all that is necceasary is to fill in a
blanks in an array job script template, which directly
runs the run_mcmc python code with appropriate options.

This requires the run_mcmc code to take command line
options.  That is easy (but not yet implemented).
The only remaining trick is to associate the task array
index with a galaxy id, within run_mcmc.  This is easily
done by globbing on the filename template and selecting
the element and parsing it for the id.

However, I've stopped working on this as making such
significant changes will require explanation and possibly
make it harder for others familiar with the old system
to run the code (in the short-term, which is what matters).

If we ever use this code for anything else, then it would
be worth completing these developments.